fix #7340 chore(project): move bugzilla into legacy
Because

* Bugzilla is only used in legacy

This commit

* Moves Bugzilla into legacy
