--- app/experimenter/bugzilla/__init__.py
@@ -1,1 +1,0 @@
-from experimenter.bugzilla.client import *  # noqa
