+++ app/experimenter/legacy/bugzilla/__init__.py
@@ -1,0 +1,1 @@
+from experimenter.legacy.bugzilla.client import *  # noqa
