+++ app/experimenter/legacy/bugzilla/tasks.py
@@ -1,8 +1,8 @@
 import markus
 from celery.utils.log import get_task_logger
 
-from experimenter.bugzilla import client as bugzilla
 from experimenter.celery import app
+from experimenter.legacy.bugzilla import client as bugzilla
 from experimenter.legacy.legacy_experiments.models import Experiment
 from experimenter.notifications.models import Notification
