+++ app/experimenter/legacy/bugzilla/tests/test_tasks.py
@@ -5,9 +5,9 @@ from markus.testing import MetricsMock
 from requests import RequestException
 
 from experimenter.base.tests.mixins import MockRequestMixin
-from experimenter.bugzilla import client as bugzilla
-from experimenter.bugzilla import tasks
-from experimenter.bugzilla.tests.mixins import MockBugzillaMixin
+from experimenter.legacy.bugzilla import client as bugzilla
+from experimenter.legacy.bugzilla import tasks
+from experimenter.legacy.bugzilla.tests.mixins import MockBugzillaMixin
 from experimenter.legacy.legacy_experiments.models import Experiment
 from experimenter.legacy.legacy_experiments.tests.factories import ExperimentFactory
 from experimenter.legacy.normandy.tests.mixins import MockNormandyMixin
