+++ app/experimenter/legacy/legacy_experiments/forms.py
@@ -10,7 +10,7 @@ from django.utils.safestring import mark_safe
 from django.utils.text import slugify
 
 from experimenter.base.models import Country, Locale
-from experimenter.bugzilla import get_bugzilla_id, tasks
+from experimenter.legacy.bugzilla import get_bugzilla_id, tasks
 from experimenter.legacy.legacy_experiments.changelog_utils import (
     ChangeLogSerializer,
     generate_change_log,
