+++ app/experimenter/legacy/legacy_experiments/tests/test_forms.py
@@ -11,7 +11,10 @@ from faker import Factory as FakerFactory
 
 from experimenter.base.tests.factories import CountryFactory, LocaleFactory
 from experimenter.base.tests.mixins import MockRequestMixin
-from experimenter.bugzilla.tests.mixins import MockBugzillaMixin, MockBugzillaTasksMixin
+from experimenter.legacy.bugzilla.tests.mixins import (
+    MockBugzillaMixin,
+    MockBugzillaTasksMixin,
+)
 from experimenter.legacy.legacy_experiments.forms import (
     RADIO_NO,
     RADIO_YES,
